Add the ability to "abort" a ConcurrentQueue.
When q.abort() is called, every active call to q.pop() wakes up
and throws "aborted". The intent is that if you have a ConcurrentQueue
that might have some waiters, you should call q.abort(), which wakes
up all the waiters, and then you yourself should wait for some kind
of acknowledgment from the waiters that it is safe to destroy q.