Implement a fire-and-forget Scheduler.
Scheduler implements schedule(), which sends a single task to the thread pool
managed by this Scheduler. This is the equivalent of N4406's async_execute(),
except that no future is returned; the model is "fire and forget".

If any task throws an exception, the behavior is undefined. (There is nobody
who could reasonably catch such an exception.)

When the Scheduler object is destroyed, it waits for tasks that are currently
in flight, and then cancels all tasks that have not yet been started.