Refactor and remove some locking from Promise.
A given SharedState is pointed to by only a few entities in the program:
one Promise and some number of Futures. Of these, only the Promise is
capable of setting ready_ to true; the Future(s) merely read the value.
Therefore, the Promise itself doesn't need to hold mtx_ when reading ready_;
no other thread can possibly be modifying ready_ at the same time.
This change removes a lock_guard from the critical path in Promise's
destructor, and allows us to express abandon_state() in terms of
set_exception().

Furthermore, the logic for setting a Promise "ready" is shared between
set_value() and set_exception(). Soon, this logic will become more
complicated; it makes sense to consolidate it into a single codepath.