Implement "UniqueFunction", the moral equivalent of std::unique_function.
UniqueFunction<void()> implements a type-erased function object
returning void. That is, it's roughly equivalent to std::function<void()>,
but has the advantage that it can capture a lambda or functor which is
non-copyable (for example, a lambda that captures a promise or future,
which is something that we'll want to do in just a little while).

UniqueFunction corresponds exactly to N4543's proposed std::unique_function,
except that only the <void()> specialization is implemented because
that's all we need right now.