Implement PackagedTask<R(A...)>.
PackagedTask is exactly analogous to std::packaged_task.
It takes a normal function or functor and "splits it up" into
the executable part (represented internally as a UniqueFunction
returning void) and the result (represented internally as a
Future<R>). The PackagedTask itself provides an operator()
for executing the executable part; the result can be retrieved
via PackagedTask::get_future().