Implement Future::then() and SharedFuture::then().
These "continuations" are not cancelable; they correspond closely
to N3721's std::future<T>::then() and std::shared_future<T>::then().
The major difference is that we don't implement "unwrapping" on the
result type: in our version,

    Future<int> f = Async([]{ return 0; })
    auto f2 = f.then([](Future<int> f){ return f; });

the type of f2 is Future<Future<int>>; whereas N3721 makes a special
case for future-returning continuations so that the type of f2 would
be simply Future<int> instead.

N3721 is unclear how the Future f is "moved to the parameter of the
continuation function". In fact, it specifically defines the result
type in terms of `decltype(func(*this))`, rather than in terms of
`decltype(func(std::move(*this)))`! I'm not sure how this is meant
to be implemented.