Implement a naive version of canceling futures.
The idea of "canceling" futures is that if you have a pending task
(launched via PackagedTask, Async, or Future::then), and you drop the
last Future that could possibly refer to the result of that task,
then the task itself should be descheduled so that its computation
doesn't uselessly tie up computing resources.

This is implemented in terms of a new member function:
Promise::has_extant_future(). Since we do not implement anything
remotely similar to a "shared promise", we can be sure that there
is only one Promise referring to a single SharedState. And there's
no way to get a Future from that Promise without calling the non-const
member function Promise::get_future(), which means that we can sensibly
implement Promise::has_extant_future() as a const member function
that simply returns whether its shared reference to the SharedState
is unique.

(Note that Folly's misleadingly named "SharedPromise" is actually
a regular "Promise" whose getFuture() method just happens to return
a SharedFuture; it is still forbidden to make copies of a Folly
SharedPromise.)

This version of canceling futures is naive in at least two ways:

- It's unconditional. You can no longer reliably "fire-and-forget" a task
via the statement `Async(mytask);`, because the future returned by Async
will be destroyed immediately. The result will be non-deterministic:
sometimes the task will execute, sometimes it will be de-scheduled.
The cancelability of a task should be determined by a parameter to
PackagedTask/Async/then; I'm not sure if it makes sense to treat
cancelability as a property of the Executor or not.

- It's excessively lazy. Dropping the last Future on a task will
cause the code of that task to get skipped when the task finally
makes it to the front of our Scheduler's task queue; but in the
meantime, the task continues to exist and hold onto resources.
For example, if the task is a lambda, its captures will stay alive
until the task reaches the front of the task queue (which means
that any captured shared_ptrs will continue to keep their controlled
resources alive). Dropping the last Future should really cause the
whole task to get destroyed immediately.

The excessive laziness is actually a correctness issue as well,
as demonstrated by test_canceling_then() in TestFuture.cc. Suppose
the programmer invokes Async(A).then(B).then(C) and then drops the
whole mess on the floor. A, B and C will be enqueued. The promise
held by C's task now !has_extant_future(), since we dropped the
result on the floor; so C will never be executed. However, the promise
held by B's task DOES has_extant_future(), because C's task holds
its future! So tasks A and B will definitely be executed, even
though B's result is used only by C and C will definitely not be
executed.