Implement "eagerly cancelling futures."
This kind of future holds a strong reference to its promise-fulfilling
task (as wrapped by PackagedTask), which means that when the last future
referencing a particular task is destroyed, the task itself is also
destroyed. Of course this extra logic applies only to futures created
by PackagedTask, or by any other caller who associates data with the
identity of the Future via SharedFuture<T>::attach_cancellable_task_state().

Note that if the state attached to a Future has a throwing destructor,
then destroying the last Future that depends on that Future will throw
that exception. For example:

    {
        auto f = Async([]{ ... });
        f = f.then([]{ ... });
        f = f.then([ capture = make_shared<ThrowOnDestruction>() ] { ... });
        f = f.then([]{ ... });
        f = f.then([]{ ... });
    }

At the closing brace, the implicit `f.~Future<int>()` will cause a chain
of destructions, leading to the destruction of `capture`, which will throw
an exception all the way back up the stack and make it appear that
`~Future<int>` has thrown an exception.

This is unfortunate "spooky action at a distance", in my opinion. But the
same thing is already true for std::shared_ptr and even std::function:
if the type-erased thing being wrapped is something that throws on
destruction, then destroying the wrapper will throw an exception.

The reason it feels strange here is that we intuitively think of a future
as something we can manipulate without executing arbitrary code; all the
code execution is wrapped up in .wait(), and is guaranteed not to "escape"
into our program until we call .get() on the future. The fact that the
future's own destructor can still execute arbitrary (even throwing) code
is surprising.