Implement WhenAll() and MakeReadyFuture().
The function WhenAll() corresponds exactly to the std::when_all()
proposed by N3784. It accepts either a (possibly empty) parameter-pack
of Futures, or else a pair of iterators denoting a range. In the
former case, it returns a Future<tuple<Future<T>...>>; in the latter
case it returns a Future<vector<Future<T>>>.

Since all of the input futures must be satisfied, it doesn't matter
which ones are satisfied first. We simply wait for the first future;
and then (literally: we attach a continuation with .then()) we wait
for the second; and so on, until all of the futures have been fulfilled.

Implementing the base case of WhenAll() requires the concept of a
"future created ready", expressed as MakeReadyFuture<R>(...), which
exactly corresponds to N3784's std::make_ready_future<R>().
N3784 also proposes std::make_exceptional_future<R>(); we don't
need that yet, so we don't implement it.